[IMP] web: convert datetime_hook tests to hoot task-id: 3705027
